:sparkles: :bug: Fixed submodules, GC & moar features
The lcmm submodule has been updated to look for the right repository.
Your unused FMix_Chunks will now properly be garbage-collected. The 
program should not segfault if it happens to still be used by the Mix 
library, as I have created a small, purposely-built garbage collector in 
C to handle these cases.
This means that your garbage chunks will be freed once they are not 
played anymore!